requirements.txt
# General requirements
numpy==1.19.1
opencv-contrib-python==3.4.2.17
scipy==1.3.0
altgraph==0.17
fbs==0.9.0
future==0.18.2
macholib==1.14
pefile==2019.4.18
PyInstaller==3.4
PyQt5==5.15.1
PyQt5-sip==12.8.1
PyYAML==5.3.1

# Segmentation
torch==1.6.0 # pytorch.org (Manually install with CUDA 10.2)
torchvision==0.7.0